Refactor navigation components - Adjusted font size of links
