+++ src/components/navigation/navigation.tsx
@@ -1,18 +1,9 @@
-import { Center, Collapse, createStyles, Group, Navbar, Stack, Text, ThemeIcon, UnstyledButton } from '@mantine/core';
-import { Link, linkCss, SubLinks } from './simple-link';
+import { GroupLink as GroupLinkComponent } from './group-link';
+import { GroupLink } from '@/types/app-types';
+import { Navbar, Stack } from '@mantine/core';
 import { SiteLogo } from '../site-logo';
-import { useRef, useState } from 'react';
 
-interface LinkGroup extends Link {
-  links?: Link[];
-  initiallyOpened?: boolean;
-}
-
-interface LinkGroupProps {
-  link: LinkGroup;
-}
-
-const links: LinkGroup[] = [
+const links: GroupLink[] = [
   { label: 'Home', url: '/', icon: 'bi-house-fill', links: [] },
   {
     label: 'Upcoming',
@@ -24,73 +15,6 @@ const links: LinkGroup[] = [
     ],
   },
 ];
-
-const useStyles = createStyles((theme) => ({
-  groupLink: {
-    ...linkCss(theme),
-    padding: '12px 16px',
-  },
-  subNav: {
-    borderLeft: '1px solid',
-    borderColor: theme.colors.gray[3],
-    marginLeft: 32,
-  },
-}));
-
-function LinkGroup({ link }: LinkGroupProps) {
-  const { classes } = useStyles();
-
-  const hasSubLinks = !!link.links?.length;
-  const [opened, setOpened] = useState(link.initiallyOpened || false);
-
-  return (
-    <>
-      <UnstyledButton
-        className={classes.groupLink}
-        sx={(theme) => ({ color: !!link.url ? theme.colors.blue[4] : undefined })}
-        onClick={() => setOpened((o) => !o)}
-      >
-        <Group>
-          {!!link.icon && (
-            <ThemeIcon p={16}>
-              <i className={`bi ${link.icon}`}></i>
-            </ThemeIcon>
-          )}
-          <Text
-            size="lg"
-            sx={{
-              flex: '1 1',
-              fontFamily: 'Montserrat, sans-serif',
-            }}
-          >
-            {link.label}
-          </Text>
-          {hasSubLinks && (
-            <Center sx={{ height: 24, width: 24 }}>
-              <i className={`bi ${opened ? 'bi-chevron-down' : 'bi-chevron-right'}`}></i>
-            </Center>
-          )}
-        </Group>
-      </UnstyledButton>
-      {hasSubLinks && (
-        <Collapse in={opened}>
-          <Stack
-            spacing={0}
-            justify="center"
-            className={classes.subNav}
-          >
-            {(link.links ?? []).map((link, index) => (
-              <SubLinks
-                key={index}
-                link={link}
-              />
-            ))}
-          </Stack>
-        </Collapse>
-      )}
-    </>
-  );
-}
 
 export function Navigation() {
   return (
@@ -107,7 +31,7 @@ export function Navigation() {
           spacing={0}
         >
           {links.map((link, index) => (
-            <LinkGroup
+            <GroupLinkComponent
               key={index}
               link={link}
             />
