+++ src/types/app-types.ts
@@ -1,0 +1,10 @@
+export interface Link {
+  label: string;
+  url?: string;
+  icon?: string;
+}
+
+export interface GroupLink extends Link {
+  links?: Link[];
+  initiallyOpened?: boolean;
+}
